requirements.txt
torch
numpy
scipy
pandas
pillow
jupyter
torchvision
pycocotools
matplotlib